requirements.txt
langchain
langgraph
langchain-community
faiss-cpu
python-dotenv